requirements.txt
zhipuai>=2.1.5
pyperclip>=1.8.2
PyQt5>=5.15.10
pytesseract>=0.3.10
Pillow>=10.0.0
pywin32>=306
